+++ c_impl/include/detectors.h
@@ -3,8 +3,9 @@
  *   FilterConvergenceAnalyzer    (aec.py 2369-2435)
  *   DoubleTalkAnalyzer           (aec.py 2437-2505)
  *
- * All three operate purely on Python `float` (fp64) scalars + bool, so C
- * mirrors with `double` and `int` (0/1).
+ * All three operate purely on float32 scalars + bool, so C mirrors with
+ * `float` and `int` (0/1). float32-by-design — the original Python fp64
+ * bit-exact parity target is retired; the drift this introduces is accepted.
  */
 #ifndef AEC_DETECTORS_H
 #define AEC_DETECTORS_H
@@ -15,14 +16,14 @@ extern "C" {
 
 /* ── RenderActivityDetector ────────────────────────────────── */
 typedef struct RenderActivity {
-    double env_mean;
-    double env_var;
+    float  env_mean;
+    float  env_var;
     int    active_prev;
     int    is_stationary;
 } RenderActivity;
 
 typedef struct RenderActivityResult {
-    double far_pwr;
+    float  far_pwr;
     int    is_active;
     int    is_stationary;
     int    warmup_active;
@@ -38,40 +39,40 @@ typedef struct FilterConvergence {
     int    converged;
     int    once_converged;
     int    conv_counter;
-    double divergence;
+    float  divergence;
 } FilterConvergence;
 
 void filter_convergence_init(FilterConvergence* c);
 void filter_convergence_reset(FilterConvergence* c);
 void filter_convergence_mark_diverged(FilterConvergence* c);
 void filter_convergence_update_divergence(FilterConvergence* c,
-                                             double near_power,
-                                             double raw_error_power);
+                                             float near_power,
+                                             float raw_error_power);
 /* Returns 1 on the transition frame. */
 int  filter_convergence_update_convergence(FilterConvergence* c,
-                                              double near_power,
-                                              double raw_error_power,
+                                              float near_power,
+                                              float raw_error_power,
                                               int far_active, int warmup_done);
 
 /* ── DoubleTalkAnalyzer ────────────────────────────────────── */
 typedef struct DoubleTalk {
-    double dt_from_energy;
-    double dt_from_shadow;
-    double shadow_advantage;
+    float  dt_from_energy;
+    float  dt_from_shadow;
+    float  shadow_advantage;
     /* Config slice */
-    double shadow_dtd_offset;            /* default 1.5 */
-    double shadow_dtd_advantage_scale;   /* default 3.0 */
+    float  shadow_dtd_offset;            /* default 1.5 */
+    float  shadow_dtd_advantage_scale;   /* default 3.0 */
 } DoubleTalk;
 
-void doubletalk_init(DoubleTalk* d, double offset, double advantage_scale);
+void doubletalk_init(DoubleTalk* d, float offset, float advantage_scale);
 void doubletalk_reset(DoubleTalk* d);
 void doubletalk_update_shadow_dt(DoubleTalk* d,
                                     int shadow_frame_count, int far_excited,
-                                    double main_err_smooth,
-                                    double shadow_err_smooth);
+                                    float main_err_smooth,
+                                    float shadow_err_smooth);
 void doubletalk_update_energy_dt(DoubleTalk* d,
-                                    int far_active, double far_pwr,
-                                    double mic_pwr, double erl_estimate);
+                                    int far_active, float far_pwr,
+                                    float mic_pwr, float erl_estimate);
 
 /* ── FilterPlateauDetector (v3.10.0 + v3.10.3 F4/H3) ─────── */
 /* Detects filter stuck below ERLE convergence threshold for sustained
@@ -80,9 +81,9 @@ void doubletalk_update_energy_dt(DoubleTalk* d,
 typedef struct FilterPlateauDetector {
     /* Config */
     int    grace_frames;        /* 400 */
-    double erle_max_db;         /* 6.0 */
-    double far_active_ratio;    /* 0.5 */
-    double dt_signal_ratio;     /* 0.10 */
+    float  erle_max_db;         /* 6.0 */
+    float  far_active_ratio;    /* 0.5 */
+    float  dt_signal_ratio;     /* 0.10 */
     int    max_attempts;        /* 2 */
     /* State */
     int    frame_count;
@@ -103,7 +104,7 @@ void filter_plateau_reset(FilterPlateauDetector* p);
 int  filter_plateau_update(FilterPlateauDetector* p,
                               int    far_active,
                               int    dt_signal_present,
-                              double erle_windowed_db,
+                              float  erle_windowed_db,
                               int    once_converged);
 
 #ifdef __cplusplus
